# sira-backend-master 2/swagger-api.json
{
    "openapi": "3.0.0",
    "paths": {
...
        "/units": {
            "post": {
                "operationId": "UnitsController_create",
                "summary": "Create a new unit (not marketing)",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateUnitDto"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "get": {
                "operationId": "UnitsController_findAll",
                "summary": "List units — sales: own uploads + published inventory only; other inventory roles: full list; marketing: full list",
                "parameters": [
                    {
                        "name": "q",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "salesId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "teamId",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/units/{id}": {
            "get": {
                "operationId": "UnitsController_findOne",
                "summary": "Get a unit by ID (sales: own or published only)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "patch": {
                "operationId": "UnitsController_update",
                "summary": "Update unit inventory (not marketing)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateUnitDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            },
            "delete": {
                "operationId": "UnitsController_remove",
                "summary": "Delete a unit (not marketing)",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/units/{id}/publish": {
            "patch": {
                "operationId": "UnitsController_publish",
                "summary": "Publish / unpublish unit and set public listing link",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/PublishUnitDto"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Units"
                ],
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
        "/webhooks/meta": {
            "get": {
                "operationId": "WebhooksController_verifyMeta",
                "summary": "Meta Webhook Verification (Challenge)",
                "parameters": [
                    {
                        "name": "hub.mode",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hub.verify_token",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "hub.challenge",
                        "required": true,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Webhooks"
                ]
            },
            "post": {
                "operationId": "WebhooksController_handleMeta",
                "summary": "Meta Lead Capture Webhook",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": ""
                    }
                },
                "tags": [
                    "Webhooks"
                ]
            }
        },
...
    },
    "info": {
        "title": "CRM API",
        "description": "Comprehensive CRM API for real estate management",
        "version": "1.0",
        "contact": {}
    },
...
    "servers": [],
    "components": {
        "securitySchemes": {
            "bearer": {
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "type": "http"
            }
        },
        "schemas": {
...
            "CreateUnitDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "Omit to auto-generate a unique code (recommended)",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "description": "Short unit / listing address line"
                    },
                    "projectName": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "monthlyInstallment": {
                        "type": "number",
                        "description": "Monthly installment"
                    },
                    "deliveryDate": {
                        "type": "string",
                        "description": "Delivery / handover date (YYYY-MM-DD)"
                    },
                    "bedrooms": {
                        "type": "number"
                    },
                    "bathrooms": {
                        "type": "number"
                    },
                    "floor": {
                        "type": "number",
                        "description": "Floor number (building level)"
                    },
                    "area": {
                        "type": "number",
                        "description": "Area in sqm or sqft"
                    },
                    "unitType": {
                        "type": "string"
                    },
                    "driveMediaLink": {
                        "type": "string",
                        "description": "Google Drive (or similar) link to media / documents — paste share link",
                        "maxLength": 1024
                    },
                    "amenities": {
                        "description": "Amenity labels (e.g. tags)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "externalLinks": {
                        "description": "External URLs (listings, walkthroughs, etc.)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "reserved",
                            "sold",
                            "unavailable"
                        ],
                        "default": "available"
                    }
                },
                "required": [
                    "description"
                ]
            },
            "PublishUnitDto": {
                "type": "object",
                "properties": {
                    "isPublished": {
                        "type": "boolean",
                        "description": "Whether the unit listing is published externally"
                    },
                    "publishedLink": {
                        "type": "string",
                        "description": "Published link payload. Supports plain URL or JSON string like {\"places\":[{\"name\":\"Facebook\",\"link\":\"https://...\"}]}",
                        "example": "{\"places\":[{\"name\":\"Facebook\",\"link\":\"https://facebook.com/...\"}]}"
                    }
                },
                "required": [
                    "isPublished"
                ]
            },
            "UpdateUnitDto": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string",
                        "description": "Omit to auto-generate a unique code (recommended)",
                        "maxLength": 100
                    },
                    "description": {
                        "type": "string"
                    },
                    "address": {
                        "type": "string",
                        "description": "Short unit / listing address line"
                    },
                    "projectName": {
                        "type": "string"
                    },
                    "location": {
                        "type": "string"
                    },
                    "price": {
                        "type": "number"
                    },
                    "monthlyInstallment": {
                        "type": "number",
                        "description": "Monthly installment"
                    },
                    "deliveryDate": {
                        "type": "string",
                        "description": "Delivery / handover date (YYYY-MM-DD)"
                    },
                    "bedrooms": {
                        "type": "number"
                    },
                    "bathrooms": {
                        "type": "number"
                    },
                    "floor": {
                        "type": "number",
                        "description": "Floor number (building level)"
                    },
                    "area": {
                        "type": "number",
                        "description": "Area in sqm or sqft"
                    },
                    "unitType": {
                        "type": "string"
                    },
                    "driveMediaLink": {
                        "type": "string",
                        "description": "Google Drive (or similar) link to media / documents — paste share link",
                        "maxLength": 1024
                    },
                    "amenities": {
                        "description": "Amenity labels (e.g. tags)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "externalLinks": {
                        "description": "External URLs (listings, walkthroughs, etc.)",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "available",
                            "reserved",
                            "sold",
                            "unavailable"
                        ],
                        "default": "available"
                    }
                }
            },
...
        }
    }
}